+++ Excel_PRIME.Tests/BugTesting.cs
@@ -1,41 +1,16 @@
-﻿using System;
-using System.Diagnostics.CodeAnalysis;
+﻿using System.Diagnostics.CodeAnalysis;
 using System.Threading.Tasks;
-
-using JetBrains.dotMemoryUnit;
 
 using NUnit.Framework;
 
 
 namespace ExcelPRIME.Tests;
 
-[ExcludeFromCodeCoverage]
-internal class DotMemoryUnit : IDisposable
-{
-    public static IDisposable Support => new DotMemoryUnit();
-
-    private DotMemoryUnit()
-    {
-        DotMemoryUnitController.TestStart();
-    }
-
-    public void Dispose() => DotMemoryUnitController.TestEnd();
-}
 
 [ExcludeFromCodeCoverage]
 [NonParallelizable]
 internal class BugTesting
 {
-    [Test]
-    [Explicit("Run DotMemory")]
-    [DotMemoryUnit(CollectAllocations = true, FailIfRunWithoutSupport=true, Directory = @".\DotMemory")]
-
-    public void Bug_000_SharedStrings_DotMemory()
-    {
-        using IDisposable dms = DotMemoryUnit.Support;
-        Bug_001_SharedStrings().GetAwaiter().GetResult();
-    }
-
     [Test]
     public async Task Bug_001_SharedStrings()
     {